app/backend/requirements.txt
astroid==2.4.2
attrs==20.3.0
autopep8==1.5.4
blis==0.7.2
catalogue==1.0.0
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cryptography==3.2.1
cycler==0.10.0
cymem==2.0.4
dnspython==1.16.0
docx2txt==0.8
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
Flask==1.1.2
Flask-Cors==3.0.9
idna==2.10
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.17.0
jsonschema==3.2.0
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.3.2
mccabe==0.6.1
murmurhash==1.0.4
nltk==3.5
numpy==1.19.4
pandas==1.1.4
pdfminer.six==20201018
pdfminer3==2018.12.3.0
Pillow==8.0.1
plac==1.1.3
preshed==3.0.4
protobuf==3.13.0
pycodestyle==2.6.0
pycparser==2.20
pycryptodome==3.9.9
pylint==2.6.0
pymongo==3.11.0
pyparsing==2.4.7
pyresparser==1.0.6
pyrsistent==0.17.3
pytesseract==0.3.6
python-dateutil==2.8.1
pytz==2020.4
regex==2020.11.13
requests==2.25.0
scikit-learn==0.23.2
scipy==1.5.4
seaborn==0.11.0
six==1.15.0
sortedcontainers==2.3.0
spacy==2.3.2
srsly==2.3.2
thinc==7.4.2
threadpoolctl==2.1.0
toml==0.10.2
tqdm==4.51.0
urllib3==1.26.2
wasabi==0.8.0
Werkzeug==1.0.1
wrapt==1.12.1
